requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1

sqlalchemy==2.0.30
asyncpg==0.29.0

alembic==1.13.1

httpx==0.27.0
python-dotenv==1.0.1
python-multipart==0.0.9

structlog==24.1.0


